Moved controllers to a separate directory and renamed insertContent to insert.
